Create hometask_airiin
Here i have the first part of the program it is a simple class that read a file and creates a structure where there is words in spanish and estonian,  there is now some simple menu where i use cero as an exit point to the program.  Next is to extract words from  the file, verify answers and start giving a score.